backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
python-multipart>=0.0.6
# Worker: PDF, OCR, inpainting, PPT
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-pptx>=0.6.21
# PaddleOCR (CPU; for GPU use paddlepaddle-gpu)
paddlepaddle>=2.5.0
paddleocr>=2.7.0
# Diffusers inpainting
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
# Celery
celery[redis]>=5.3.0
